+++ src/pages/Dashboard/Hotel/index.js
@@ -1,3 +1,17 @@
+import { useEffect, useState } from 'react';
+import useEnrollment from '../../../hooks/api/useEnrollment';
 export default function Hotel() {
-  return 'Hotel: Em breve!';
+  // const { enrollment, getEnrollment } = useEnrollment();
+  // const [Enroll, SetEnroll] = useState('');
+  // useEffect(() => {
+  //   getEnroll();
+  // }, []);
+
+  // async function getEnroll() {
+  //   const enrollapi = await getEnrollment();
+  //   SetEnroll(enrollapi.name); 
+  // }
+  return (
+    <p>Oi</p>
+  );
 }
